Updated rlRandomName with better debugging and error message
Removed pdb.set_trace, as the Python interpreter in VS Code threw up a warning for possible incompatibility with its own debugger. Replaced it with a very descriptive error message.

Also fixed rlRandomName's behavior when LogicController uses it as a module. Not useful except for debugging, but still good to have.