FIX (CLI): Fix two unreported bugs in WorseCLI submenus
Bug A: Trying to exit from WorseCLI submenus wouldn't work because that WorseCLI.py file didn't have 'sys' imported. Fixed.

Bug B: When the user tries to generate a list of randoms names, the program prompts them for the desired number of names to generate. If the user responded with a non-integer, an uncaught ValueError exception was raised. The program now catches the exception, informs the user of the error, and prompts them again for the desired number of names.