FIX: Change import statements to lowercase
The 'rosevomit/programlogic/__init__.py' file would cause the program to crash because its import statements still referred to the modules by the names they had prior to my "change-all-the-module-names-to-lowercase" editing spree. This is now fixed.

A similar problem existed in the 'devtests/__init__.py' file, though this didn't actually cause any problems because it was never called by anything. In that case, I simply removed the import statements entirely. We'll add back the correct ones if needed in the future.