FEAT (settings): Add ability to change settings within CLI
This commit adds the ability for the user to change settings without having to leave the program and manually changing the .ini file. (Issue #30)

Most of the work is in 'core/settings', although I did change some function names that I then had to change whenever they were called elsewhere in the program ('programcli/worsecli', 'core/utilities').

Additionally, we were importing core.settings in core/startup and never using it, so I deleted that import statement.