FEAT (logging): Log according to settings-data.ini (+ comments)
This commit changes core.settings.restore_file() so that it includes helpful comments when regenerating the settings-data.ini file. 

This commit also changes the behavior of core.logs.start_logging() so that output files are not created when core.settings.logging_service() returns False.

This commit is part of ongoing work on issue #27.