DOCS (docstrings): Add NumPy-style docstrings everywhere
Issue #44, standardizing docstrings. This commit is a big "first pass" at detailed, NumPy-style docstrings. I'm just trying them out - I'm still not used to them. 
 
Skipped a lot in programcli.worsecli and programcli.bettercli because the CLI redo is coming in the next version and a lot will change anyway. Skipped some stuff in programlogic.suncalc, because that's due for a big rewrite anyways before the release of this version.